All Checkboxes are checked as far as I know!

+++ server/src/services/TicketsService.js
@@ -5,7 +5,8 @@ import { eventsService } from "./EventsService.js"
 class TicketsService {
     async createTicket(ticketData) {
         const event = await eventsService.getEventById(ticketData.eventId)
-
+        event.capacity--
+        event.save()
         if (event.isCanceled) {
             throw new Forbidden(`${event.name} Has been canceled sorry for inconvenience`)
         }
@@ -19,6 +20,9 @@ class TicketsService {
 
     async refundTicket(ticketId, userId) {
         const ticket = await dbContext.Tickets.findById(ticketId)
+        if (ticket.accountId.toString() != userId) {
+            throw new Forbidden("not your ticket to delete")
+        }
         await ticket.delete()
         return "Ticket Refunded!"
     }
